dashboard: align footer divider with the rail 'Report a problem' divider
The footer (.cx-foot-bar) separator sat ~7px above the left-rail
.cx-nav-aux separator, so the two 1px lines stepped at the rail/content
boundary. Both are bottom-anchored; trimming the footer's vertical
padding (s16->s8) drops its top border to the same baseline as the rail
divider (and stops the footer overflowing below the fold).

+++ cortex_web/apps/web/src/components/Shell.tsx
@@ -230,7 +230,7 @@ td.ellcell .of{color:var(--ink-faint);}
 .cx-deckhint .k{display:inline-flex;align-items:center;gap:4px;margin-right:var(--s12);}
 .cx-deckhint .k i{width:8px;height:8px;border-radius:0;}
 
-.cx-foot-bar{padding:var(--s16) var(--s32);border-top:1px solid var(--bd-subtle);
+.cx-foot-bar{padding:var(--s8) var(--s32);border-top:1px solid var(--bd-subtle);
   display:flex;align-items:center;justify-content:space-between;gap:var(--s16);flex-wrap:wrap;}
 .cx-foot-bar a{font-size:12px;font-weight:500;color:var(--ink-subtle);text-decoration:none;}
 .cx-foot-bar a:hover{color:var(--ink);text-decoration:underline;}
